progress: clean up message components
